refactor: 회원가입 페이지 리팩토링
- RegisterClient, GeneralForm, DojangForm 컴포넌트 분리
- Field, InputWithIcon → src/components/common으로 분리
- BeltSelect, NicknameField, BusinessFileUpload → register/components로 분리
- useNicknameCheck → src/hooks로 분리 및 useDebounce 적용
- 닉네임 자동 중복확인으로 변경 (useDebounce 500ms)
- 비밀번호 정규식 추가 (8자 이상 + 영문 + 숫자 + 특수문자)
- 사업자등록번호 숫자만 입력 가능하도록 정규식 추가
- 연락처 형식 검사 추가
- BeltSelect react-hook-form 연동 수정
- role="alert" 접근성 적용
- role="tab", aria-selected 탭 접근성 적용
- aria-busy 버튼 적용
- robots에 follow: false 추가
- 불필요한 supabase import 제거
