+++ src/components/common/Field.tsx
@@ -1,0 +1,21 @@
+export default function Field({
+  label,
+  htmlFor,
+  children,
+}: {
+  label: string;
+  htmlFor: string;
+  children: React.ReactNode;
+}) {
+  return (
+    <div>
+      <label
+        htmlFor={htmlFor}
+        className="block text-sm font-medium text-text-primary mb-2"
+      >
+        {label}
+      </label>
+      {children}
+    </div>
+  );
+}
